Time: 19 ms (8.85%) | Memory: 26.2 MB (15.73%) - LeetSync

+++ 1408-find-the-smallest-divisor-given-a-threshold/find-the-smallest-divisor-given-a-threshold.cpp
@@ -5,7 +5,7 @@ public:
         while (low < high) {
             mid = low + (high - low) / 2, sum = 0;
             for (int i=0;i<nums.size();i++) {
-                sum += ceil(nums[i] / double(mid));
+                sum += ceil(nums[i] / float(mid));
                 if (sum > threshold) {
                     break;
                 }
